+ Админпанель заготовка

+++ index.php
@@ -7,16 +7,17 @@
 	require __DIR__ . '/autoload.php';
 	
 	
-	//$user = User::findById(6);
-	/*
+	$user = User::findById(3);
 	
+	/*
 	$user->name = 'Vasiliska';
 	$user->email = 'test1@test.com';
-	
-	$user->delete();
-	$users = User::findAll();
-	var_dump($users);
 	*/
+	if ($user){
+		$user->delete();
+	}
+	$users = User::findAll();
+	
 	$news = News::getByCount(3);
 	include __DIR__ . '/App/templates/news.php';
